chore(e2e): refactor breakdowns
